feat: revise model, view-all gudang, and detail gudang

+++ silogistik2106637555/src/main/java/apap/ti/silogistik2106637555/dto/request/CreateBarangRequestDTO.java
@@ -12,5 +12,5 @@ import lombok.NoArgsConstructor;
 public class CreateBarangRequestDTO {
     private String merk;
     private int tipeBarang;
-    private BigInteger hargaBarang;
+    private long hargaBarang;
 }
